chore(stackable-versioned): Move K8s related error enums
