+++ crates/stackable-versioned/src/k8s.rs
@@ -1,12 +1,16 @@
 use std::collections::HashMap;
 
 use k8s_version::Version;
+#[cfg(doc)]
+use kube::core::conversion::ConversionReview;
 use schemars::schema::{InstanceType, Schema, SchemaObject, SingleOrVec};
+use snafu::Snafu;
 
 // NOTE (@Techassi): This struct represents a rough first draft of how tracking values across
 // CRD versions can be achieved. It is currently untested and unproven and might change down the
 // line. Currently, this struct is only generated by the macro but not actually used by any other
 // code. The tracking itself will be introduced in a follow-up PR.
+/// Contains changed values during upgrades and downgrades of CRDs.
 #[derive(Clone, Debug, serde::Deserialize, serde::Serialize, schemars::JsonSchema)]
 pub struct ChangedValues {
     /// List of values needed when downgrading to a particular version.
@@ -16,6 +20,7 @@ pub struct ChangedValues {
     pub upgrades: HashMap<Version, Vec<ChangedValue>>,
 }
 
+/// Contains a changed value for a single field of the CRD.
 #[derive(Clone, Debug, serde::Deserialize, serde::Serialize, schemars::JsonSchema)]
 pub struct ChangedValue {
     /// The name of the field of the custom resource this value is for.
@@ -39,4 +44,31 @@ fn raw_object_schema(_: &mut schemars::r#gen::SchemaGenerator) -> Schema {
         .into(),
         ..Default::default()
     })
+}
+
+/// This error indicates that parsing an object from a [`ConversionReview`] failed.
+#[derive(Debug, Snafu)]
+pub enum ParseObjectError {
+    #[snafu(display(r#"failed to find "apiVersion" field"#))]
+    FieldNotPresent,
+
+    #[snafu(display(r#"the "apiVersion" field is not a string"#))]
+    FieldNotStr,
+
+    #[snafu(display("encountered unknown object api version {api_version:?}"))]
+    UnknownApiVersion { api_version: String },
+
+    #[snafu(display("failed to deserialize object from json"))]
+    Deserialize { source: serde_json::Error },
+}
+
+/// This error indicates that converting an object from a [`ConversionReview`] to the desired
+/// version failed.
+#[derive(Debug, Snafu)]
+pub enum ConvertObjectError {
+    #[snafu(display("failed to parse object"))]
+    Parse { source: ParseObjectError },
+
+    #[snafu(display("failed to serialize object into json"))]
+    Serialize { source: serde_json::Error },
 }
